requirements.txt
absl-py==0.10.0
numpy==1.19.1
Pillow==7.2.0
sentencepiece==0.1.92
six==1.15.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
tensorflow-hub==0.9.0


